--- models/src/main/kotlin/io/provenance/api/models/p8e/tx/permissions/fees/FeeGrant.kt
@@ -1,7 +1,0 @@
-package io.provenance.api.models.p8e.tx.permissions.fees
-
-data class FeeGrant(
-    val granter: String,
-    val grantee: String,
-    val allowance: Allowance?
-)
