+++ models/src/main/kotlin/io/provenance/api/models/p8e/tx/permissions/fees/grant/FeeGrantDetails.kt
@@ -1,0 +1,8 @@
+package io.provenance.api.models.p8e.tx.permissions.fees.grant
+
+import io.provenance.api.models.p8e.tx.permissions.fees.Allowance
+
+data class FeeGrantDetails(
+    val grantee: String,
+    val allowance: Allowance
+)
