+++ service/src/main/kotlin/io/provenance/api/domain/usecase/provenance/tx/permissions/fees/get/GetFeeGrant.kt
@@ -4,7 +4,7 @@ import io.provenance.api.domain.usecase.AbstractUseCase
 import io.provenance.api.domain.usecase.common.errors.NotFoundError
 import io.provenance.api.domain.usecase.provenance.tx.permissions.fees.get.models.GetFeeGrantAllowanceRequestWrapper
 import io.provenance.api.frameworks.provenance.extensions.toModel
-import io.provenance.api.models.p8e.tx.permissions.fees.FeeGrant
+import io.provenance.api.models.p8e.tx.permissions.fees.get.GetFeeGrantResponse
 import io.provenance.client.grpc.GasEstimationMethod
 import io.provenance.client.grpc.PbClient
 import io.provenance.client.protobuf.extensions.getFeeGrant
@@ -12,7 +12,7 @@ import java.net.URI
 import org.springframework.stereotype.Component
 
 @Component
-class GetFeeGrant : AbstractUseCase<GetFeeGrantAllowanceRequestWrapper, FeeGrant>() {
+class GetFeeGrant : AbstractUseCase<GetFeeGrantAllowanceRequestWrapper, GetFeeGrantResponse>() {
     override suspend fun execute(args: GetFeeGrantAllowanceRequestWrapper) =
         PbClient(args.request.chainId, URI(args.request.nodeEndpoint), GasEstimationMethod.MSG_FEE_CALCULATION).use { pbClient ->
             runCatching {
